backend/requirements.txt
# Backend dependencies for AI Document Q&A Application

# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
pytesseract==0.3.10
pdf2image==1.17.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Database
motor==3.3.2
pymongo==4.6.1
redis==5.0.1

# AI/ML
langchain>=0.1.6
langchain-community>=0.0.19
langchain-huggingface>=0.0.1
sentence-transformers>=2.6.0
faiss-cpu>=1.7.4
huggingface-hub>=0.23.0
transformers>=4.39.0

# Audio/Video Processing
openai-whisper==20231117
ffmpeg-python==0.2.0
pydub==0.25.1

# PDF Processing
pymupdf==1.23.21
pdfplumber==0.10.4

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator>=2.0.0
aiofiles==23.2.1
httpx==0.26.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.23
pytest-cov>=4.1.0
pytest-mock>=3.12.0
